se arreglaron plugins del pom y se creo un api para la clase productos
