+++ src/main/java/co/edu/eci/cvds/controller/ProductController.java
@@ -8,43 +8,11 @@ import co.edu.eci.cvds.service.ProductService;
 
 import java.util.List;
 
-@RestController
-@RequestMapping("/api/products") 
+@Controller
+@RequestMapping("/products") 
 public class ProductController { 
 
-    @Autowired
-    private ProductService productService; 
-
-    @GetMapping
-    public List<Product> getAllProducts() { 
-        return productService.getAllProducts();
-    }
-
-    @GetMapping("/searchByName/{name}") 
-    public List<Product> searchProductsByName(@PathVariable String name) { 
-        return productService.getProductsByName(name);
-    }
-
-    @GetMapping("/{id}")
-    public Product getProductById(@PathVariable String id) {
-        return productService.getProductById(id);
-    }
-
-    @PostMapping
-    public Product createProduct(@RequestBody Product product) {
-        return productService.createProduct(product);
-    }
-
-    @DeleteMapping("/{id}")
-    public void deleteProduct(@PathVariable String id) {
-        productService.deleteProduct(id);
-    }
-
-    @PutMapping("/{id}")
-    public Product updateProduct(@PathVariable String id, @RequestBody Product product) {
-        return productService.updateProduct(id, product);
-    }
-
+    
     @GetMapping("/productos") 
     public String showProductosPage() {
         return "productos";
